api url fix+few more
