+++ app/client/cypress/e2e/Regression/ClientSide/BugTests/AbortAction_Spec.ts
@@ -10,7 +10,8 @@ import {
   createMessage,
 } from "../../../../support/Objects/CommonErrorMessages";
 
-const largeResponseApiUrl = "https://api.publicapis.org/entries";
+const largeResponseApiUrl = "https://api.github.com/emojis";
+//"https://api.publicapis.org/entries";
 //"https://jsonplaceholder.typicode.com/photos";//Commenting since this is faster sometimes & case is failing
 
 describe("Abort Action Execution", function () {
