+++ app/client/cypress/limited-tests.txt
@@ -1,5 +1,7 @@
 # To run only limited tests - give the spec names in below format:
-cypress/e2e/Regression/ClientSide/Templates/Fork_Template_spec.js
+#cypress/e2e/Regression/ClientSide/Templates/Fork_Template_spec.js
+cypress/e2e/Regression/ClientSide/Widgets/ListV2/DataIdentifierProperty_spec.ts
+cypress/e2e/Regression/ClientSide/BugTests/AbortAction_Spec.ts
 
 
 # For running all specs - uncomment below:
